restructure unit tests
