+++ tests/test_op_individual.py
@@ -1,6 +1,8 @@
-
+import pandas as pd
 from unittest import TestCase
-from oncopacket.model import OpIndividual
+import phenopackets as PPkt
+from src.oncopacket.model import OpIndividual
+from src.oncopacket.cda import CdaIndividualFactory
 
 class OpIndividualTestCase(TestCase):
 
@@ -23,21 +25,39 @@ class OpIndividualTestCase(TestCase):
         cls._series = pd.Series(data, index=column_names)
 
     def test_creation(self):
-        ifact = Op()
+        ifact = CdaIndividualFactory()
         self.assertIsNotNone(ifact)
         ga4gh_indi = ifact.from_cancer_data_aggregator(self._series)
         self.assertIsNotNone(ga4gh_indi)
 
     def test_id(self):
-        ifact = IndividualFactory()
+        ifact = CdaIndividualFactory()
         ga4gh_indi = ifact.from_cancer_data_aggregator(self._series)
         expected_id = 'CGCI.HTMCP-03-06-02007'
         self.assertEquals(expected_id, ga4gh_indi.id)
 
     def test_iso_age(self):
-        ifact = IndividualFactory()
+        """
+        TODO - iso conversion not working (e.g. 23M
+        """
+        ifact = CdaIndividualFactory()
         ga4gh_indi = ifact.from_cancer_data_aggregator(self._series)
-        expected_iso = 'P43Y43M6D'
+        expected_iso = 'P43Y23M6D'
         calculated_iso = ga4gh_indi.time_at_last_encounter.age.iso8601duration
         self.assertEquals(expected_iso, calculated_iso)
+
+    def test_taxonomy(self):
+        ifact = CdaIndividualFactory()
+        ga4gh_indi = ifact.from_cancer_data_aggregator(self._series)
+        taxonomy = ga4gh_indi.taxonomy
+        self.assertIsNotNone(taxonomy)
+        self.assertEquals("NCBITaxon:9606", taxonomy.id)
+        self.assertEquals("homo sapiens sapiens", taxonomy.label)
+
+    def test_alive_vital_status(self):
+        ifact = CdaIndividualFactory()
+        ga4gh_indi = ifact.from_cancer_data_aggregator(self._series)
+        vs = ga4gh_indi.vital_status
+        self.assertIsNotNone(vs)
+        self.assertEquals(PPkt.VitalStatus.ALIVE, vs.status)
